This PR is a long list of changes suggested by clang-tidy, or inspired
by some of the feedback.

While this represents a lot of churn, it also allows us to enable many
clang-tidy checks, and hopefully keep a more consistent style now.

Performance, I believe is unaffected:
https://rebench.dev/SOMpp/compare/dc3c3acd1c15f1f4f9dfb41e3ed09d5b5c0a4cdf..b4f880fd6e096dfed57b1daa2b6dcf42f85824a0

+++ src/compiler/BytecodeGenerator.h
@@ -33,16 +33,16 @@
 #include "MethodGenerationContext.h"
 
 void Emit1(MethodGenerationContext& mgenc, uint8_t bytecode,
-           size_t stackEffect);
+           int64_t stackEffect);
 void Emit2(MethodGenerationContext& mgenc, uint8_t bytecode, size_t idx,
-           size_t stackEffect);
+           int64_t stackEffect);
 void Emit3(MethodGenerationContext& mgenc, uint8_t bytecode, size_t idx,
-           size_t ctx, size_t stackEffect);
+           size_t ctx, int64_t stackEffect);
 
 void EmitHALT(MethodGenerationContext& mgenc);
 void EmitDUP(MethodGenerationContext& mgenc);
-void EmitPUSHLOCAL(MethodGenerationContext& mgenc, long idx, int ctx);
-void EmitPUSHARGUMENT(MethodGenerationContext& mgenc, long idx, int ctx);
+void EmitPUSHLOCAL(MethodGenerationContext& mgenc, size_t idx, size_t ctx);
+void EmitPUSHARGUMENT(MethodGenerationContext& mgenc, size_t idx, size_t ctx);
 void EmitPUSHFIELD(MethodGenerationContext& mgenc, VMSymbol* field);
 void EmitPUSHBLOCK(MethodGenerationContext& mgenc, VMInvokable* block);
 void EmitPUSHCONSTANT(MethodGenerationContext& mgenc, vm_oop_t cst);
@@ -50,8 +50,8 @@ void EmitPUSHCONSTANT(MethodGenerationContext& mgenc, uint8_t literalIndex);
 void EmitPUSHCONSTANTString(MethodGenerationContext& mgenc, VMString* str);
 void EmitPUSHGLOBAL(MethodGenerationContext& mgenc, VMSymbol* global);
 void EmitPOP(MethodGenerationContext& mgenc);
-void EmitPOPLOCAL(MethodGenerationContext& mgenc, long idx, int ctx);
-void EmitPOPARGUMENT(MethodGenerationContext& mgenc, long idx, int ctx);
+void EmitPOPLOCAL(MethodGenerationContext& mgenc, size_t idx, size_t ctx);
+void EmitPOPARGUMENT(MethodGenerationContext& mgenc, size_t idx, size_t ctx);
 void EmitPOPFIELD(MethodGenerationContext& mgenc, VMSymbol* field);
 void EmitSEND(MethodGenerationContext& mgenc, VMSymbol* msg);
 void EmitSUPERSEND(MethodGenerationContext& mgenc, VMSymbol* msg);
@@ -73,7 +73,7 @@ size_t EmitJumpIfGreaterWithDummyOffset(MethodGenerationContext& mgenc);
 void EmitJumpBackwardWithOffset(MethodGenerationContext& mgenc,
                                 size_t jumpOffset);
 size_t Emit3WithDummy(MethodGenerationContext& mgenc, uint8_t bytecode,
-                      size_t stackEffect);
+                      int64_t stackEffect);
 
 void EmitPushFieldWithIndex(MethodGenerationContext& mgenc, uint8_t fieldIdx);
 void EmitPopFieldWithIndex(MethodGenerationContext& mgenc, uint8_t fieldIdx);
